refactor: finalize driver conversion to classes
- Adjusted the rest of the code based on the new driver format.
- Upper cased the exported driver constructor names from the drivers
module because that's the convention for constructors.
- Ensured backwards compatibility in the package's interface, by upper
casing or lower casing the first letter of the driver names where
necessary and making sure the underscore in front of 3xbit is ignored.
- Adjusted the unit test that checks if the drivers are added to the
index.js in alphabetical order based on the slight difference between the
file names and the keys in the index now.
- Updated the example in the README for the new format.